Refactor: RectangleViewFactory에서 리턴하는 RectangleView 사용하도록 수정
- UIView를 ViewController에서 직접 생성하지 않고 별도의 팩토리에서 생성해서 받아쓰도록 수정
- RectangleViewFactory의 메소드에서 RectangleView 타입 인스턴스 리턴
- 리턴한 해당 인스턴스를 ViewController에서 CanvasView로 넘겨서 화면에 출력
